Enabling command line and IDE based individual tests which use this resource.

+++ stack/test-utils/src/main/java/org/usergrid/cassandra/CassandraResource.java
@@ -22,36 +22,95 @@ import org.springframework.context.support.ClassPathXmlApplicationContext;
 public class CassandraResource extends ExternalResource
 {
     private static final String TMP = "tmp";
-    public static final Logger logger = LoggerFactory.getLogger( CassandraResource.class );
-
-    ConfigurableApplicationContext applicationContext;
-    CassandraDaemon cassandraDaemon;
+    public static final Logger LOG = LoggerFactory.getLogger( CassandraResource.class );
 
     private final String schemaManagerName;
+    private final Object lock = new Object();
+    private boolean initialized = false;
+    private ConfigurableApplicationContext applicationContext;
+    private CassandraDaemon cassandraDaemon;
     private SchemaManager schemaManager;
 
 
+    /**
+     * Creates a Cassandra starting ExternalResource for JUnit test cases
+     * which uses the default SchemaManager for Cassandra.
+     */
     public CassandraResource()
     {
         this( null );
     }
 
 
-    public CassandraResource(String schemaManagerName)
+    /**
+     * Creates a Cassandra starting ExternalResource for JUnit test cases
+     * which uses the specified SchemaManager for Cassandra.
+     */
+    public CassandraResource( String schemaManagerName )
     {
         this.schemaManagerName = schemaManagerName;
     }
 
 
+    /**
+     * Gets a bean from the application context.
+     *
+     * @param requiredType the type of the bean
+     * @param <T> the type of the bean
+     * @return the bean
+     */
     public <T> T getBean( String name, Class<T> requiredType )
     {
+        protect();
         return applicationContext.getBean( name, requiredType );
     }
 
 
+    /**
+     * Gets a bean from the application context.
+     *
+     * @param requiredType the type of the bean
+     * @param <T> the type of the bean
+     * @return the bean
+     */
     public <T> T getBean( Class<T> requiredType )
     {
+        protect();
         return applicationContext.getBean( requiredType );
+    }
+
+
+    /**
+     * Gets whether this resource is ready to use.
+     *
+     * @return true if ready to use, false otherwise
+     */
+    public boolean isReady()
+    {
+        return initialized;
+    }
+
+
+    /**
+     * Protects against IDE or command line runs of a unit test which when
+     * starting the test outside of the Suite will not start the resource.
+     * This makes sure the resource is automatically started on a usage
+     * attempt.
+     */
+    private void protect()
+    {
+        if ( ! isReady() )
+        {
+            try
+            {
+                before();
+            }
+            catch ( Throwable t )
+            {
+                LOG.error("Failed to start up Cassandra.", t);
+                throw new RuntimeException( t );
+            }
+        }
     }
 
 
@@ -63,23 +122,34 @@ public class CassandraResource extends ExternalResource
     @Override
     protected void before() throws Throwable
     {
-        super.before();
+        synchronized ( lock )
+        {
+            super.before();
 
-        logger.info( "Initializing Cassandra..." );
+            if ( isReady() )
+            {
+                return;
+            }
 
-        System.setProperty( "cassandra-foreground", "true" );
-        System.setProperty( "log4j.defaultInitOverride", "true" );
-        System.setProperty( "log4j.configuration", "log4j.properties" );
-        System.setProperty( "cassandra.ring_delay_ms", "100" );
+            LOG.info("Initializing Cassandra...");
 
-        FileUtils.deleteQuietly( new File( TMP ) );
+            System.setProperty( "cassandra-foreground", "true" );
+            System.setProperty( "log4j.defaultInitOverride", "true" );
+            System.setProperty( "log4j.configuration", "log4j.properties" );
+            System.setProperty( "cassandra.ring_delay_ms", "100" );
 
-        cassandraDaemon = new CassandraDaemon();
-        cassandraDaemon.activate();
-        String[] locations = { "usergrid-test-context.xml" };
-        applicationContext = new ClassPathXmlApplicationContext( locations );
+            FileUtils.deleteQuietly( new File( TMP ) );
 
-        loadSchemaManager( schemaManagerName );
+            cassandraDaemon = new CassandraDaemon();
+            cassandraDaemon.activate();
+            String[] locations = { "usergrid-test-context.xml" };
+            applicationContext = new ClassPathXmlApplicationContext( locations );
+
+            loadSchemaManager( schemaManagerName );
+            initialized = true;
+
+            lock.notifyAll();
+        }
     }
 
 
@@ -87,11 +157,11 @@ public class CassandraResource extends ExternalResource
      * Stops Cassandra after a TestSuite or test Class executes.
      */
     @Override
-    protected void after()
+    protected synchronized void after()
     {
         super.after();
 
-        logger.info( "Shutting down Cassandra" );
+        LOG.info("Shutting down Cassandra");
 
         if ( schemaManager != null )
         {
@@ -114,24 +184,29 @@ public class CassandraResource extends ExternalResource
     }
 
 
+    /**
+     * Loads the specified {@link SchemaManager} or the default manager if
+     * the manager name that is provided is null.
+     *
+     * @param schemaManagerName the name of the SchemaManager to load, or null
+     */
     private void loadSchemaManager( String schemaManagerName )
     {
         if ( ! applicationContext.isActive() )
         {
-            logger.info( "Restarting context..." );
+            LOG.info( "Restarting context..." );
             applicationContext.refresh();
         }
 
         if ( schemaManager != null )
         {
-            // TODO check for classpath and go static?
-            logger.info( "dataControl found - looking upma SchemaManager impl" );
-            this.schemaManager = getBean( schemaManagerName, SchemaManager.class );
+            LOG.info( "Looking up SchemaManager impl: {}", schemaManagerName );
+            this.schemaManager = applicationContext.getBean( schemaManagerName, SchemaManager.class );
         }
         else
         {
-            logger.info( "SchemaManager is null - using the default SchemaManager impl" );
-            this.schemaManager = getBean( SchemaManager.class );
+            LOG.info( "The SchemaManager is not specified - using the default SchemaManager impl" );
+            this.schemaManager = applicationContext.getBean( SchemaManager.class );
         }
 
         schemaManager.create();
